# backend/openapi.yaml
openapi: 3.1.0
info:
  description: |
    사용량 기반 과금 플랫폼의 API. 사용량 이벤트를 수집하고, 고객별 월 사용량과 청구 예정액을 조회한다.

    모든 요청에 X-Organization-Id 헤더로 도입사 ID를 보낸다. 인증이 붙기 전까지 쓰는 임시 헤더다.

    오류 응답은 code, message, errors로 이뤄진다. 오류별 처리는 code로 분기한다. message는
    code마다 하나인 한국어 문구이고, errors는 틀린 필드를 짚을 수 있는 400에만 실린다.
  title: MeterEngine API
  version: v1
servers:
- description: 이 문서를 제공한 호스트
  url: /
paths:
...
  /v1/events:
    get:
      description: |
        지정한 달의 이벤트를 occurred_at 최신순으로 한 페이지씩 돌려준다.
        occurred_at이 같으면 transaction_id 내림차순이다.
        기간은 KST 기준의 달이다. occurred_at 2026-08-31T23:59:59+09:00 이벤트는 8월에,
        2026-09-01T00:00:00+09:00 이벤트는 9월에 든다. 사용량 조회와 기준이 같다.
        customer_id, month, type을 함께 주면 모두 만족하는 이벤트만 나오고, total은 조건을 적용한 뒤의 건수다.
        properties는 저장된 값 그대로다. 키 순서와 공백은 보낸 것과 다를 수 있고, 숫자 자릿수는 그대로다.
      operationId: listEvents
      parameters:
      - description: 도입사 ID. 인증이 붙기 전까지 쓰는 임시 헤더다.
        in: header
        name: X-Organization-Id
        required: true
        schema:
          type: string
          format: uuid
      - description: 0부터 세는 페이지 번호.
        in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 0
          minimum: 0
      - description: 한 페이지에 담을 개수.
        in: query
        name: size
        required: false
        schema:
          type: integer
          format: int32
          default: 20
          maximum: 100
          minimum: 1
      - description: 고객을 좁힌다. 생략하면 모든 고객이 대상이다.
        in: query
        name: customer_id
        required: false
        schema:
          type: string
          format: uuid
      - description: "조회할 달(yyyy-MM, KST). 생략하면 이번 달이다."
        example: 2026-08
        in: query
        name: month
        required: false
        schema:
          type: string
      - description: 이벤트 type을 좁힌다. 어느 미터의 event_type도 아닌 값이어도 저장된 이벤트는 조회된다.
        in: query
        name: type
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ListEventsResponse"
          description: 한 페이지 분량의 이벤트. 조건에 맞는 것이 없으면 events가 빈 배열이다
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: |
            code=validation_error: page가 음수거나 size가 1~100 밖이거나, customer_id가 UUID가
            아니거나 month가 yyyy-MM이 아니거나, X-Organization-Id가 없거나 UUID가 아니다.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "code=customer_not_found: customer_id가 가리키는 고객이 없다"
      summary: 사용량 이벤트 조회
      tags:
      - event-controller
    post:
      description: |
        사용량 이벤트 한 건을 저장한다.
        transaction_id가 같은 이벤트를 다시 보내면 저장하지 않고 처음 저장한 이벤트를 유지한다.
        이때도 응답은 200이고 duplicate가 true다. 응답을 받지 못했으면 같은 transaction_id로 다시 보내면 된다.
        properties의 내용은 검증하지 않는다. 키 순서와 공백은 보존되지 않고,
        숫자는 정수와 소수 모두 자릿수가 잘리지 않는다.
        timestamp는 이벤트가 발생한 시각이고 조회 응답에서는 occurred_at으로 나온다.
        received_at은 서버가 받은 시각이며 요청에 실어도 무시된다.
      operationId: ingestEvent
      parameters:
      - description: 도입사 ID. 인증이 붙기 전까지 쓰는 임시 헤더다.
        in: header
        name: X-Organization-Id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IngestEventRequest"
        required: true
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/IngestEventResponse"
          description: 저장했다(duplicate=false). 이미 저장된 transaction_id면 duplicate=true다
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: |
            code=validation_error: 필수 필드가 없거나 형식이 틀렸다. 어느 필드가 왜 거절됐는지는 errors에 있다.
            code=invalid_event: 저장할 수 없는 값이 들어 있다. 같은 본문을 다시 보내도 성공하지 않는다.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "code=customer_not_found: customer_id가 가리키는 고객이 없다"
      summary: 사용량 이벤트 수집
      tags:
      - event-controller
  /v1/invoices/draft:
    get:
      description: |
        모든 고객의 청구 예정액을 미터별 사용량 x 단가로 계산해 돌려준다.
        조회할 때마다 계산하며 저장하지 않는다. 기간은 KST 기준의 달이다.
        금액은 원 단위 정수다. 라인(고객 x 미터)마다 원 미만을 버린 뒤 합산한다.
        이벤트가 없는 고객도 단가가 있는 미터마다 수량 0, 금액 0인 라인이 나온다.
        고객은 이름 오름차순, 라인은 미터 code 오름차순이다.
      operationId: previewDraftInvoice
      parameters:
      - description: 도입사 ID. 인증이 붙기 전까지 쓰는 임시 헤더다.
        in: header
        name: X-Organization-Id
        required: true
        schema:
          type: string
          format: uuid
      - description: "집계할 달(yyyy-MM, KST). 생략하면 이번 달이다."
        example: 2026-08
        in: query
        name: month
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/DraftInvoiceResponse"
          description: 고객별 청구 예정액. 등록한 미터가 없으면 각 고객의 lines가 빈 배열이다
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "code=validation_error: X-Organization-Id가 없거나 UUID가 아니거나,\
            \ month가 yyyy-MM이 아니다"
      summary: 청구 예정액 조회
      tags:
      - draft-invoice-controller
  /v1/usage:
    get:
      description: |
        미터마다 type이 그 미터의 event_type과 같은 이벤트를 고객별로 합산해 돌려준다.
        기간은 KST 기준의 달이다. occurred_at 2026-08-31T23:59:59+09:00 이벤트는 8월에,
        2026-09-01T00:00:00+09:00 이벤트는 9월에 든다.
        모든 고객이 응답에 나오며, 이벤트가 없는 고객은 quantity가 0이다.
        properties의 target_property 값이 숫자가 아닌 이벤트는 합계에서 빠진다.
        금액은 싣지 않는다. 금액은 청구 예정액 조회(GET /v1/invoices/draft)가 준다.
      operationId: aggregateBillableMetricUsages
      parameters:
      - description: 도입사 ID. 인증이 붙기 전까지 쓰는 임시 헤더다.
        in: header
        name: X-Organization-Id
        required: true
        schema:
          type: string
          format: uuid
      - description: "집계할 달(yyyy-MM, KST). 생략하면 이번 달이다."
        example: 2026-08
        in: query
        name: month
        required: false
        schema:
          type: string
      responses:
        "200":
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ListBillableMetricUsagesResponse"
          description: "미터별, 고객별 사용량. 등록한 미터가 없으면 billable_metric_usages가 빈 배열이다"
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: "code=validation_error: X-Organization-Id가 없거나 UUID가 아니거나,\
            \ month가 yyyy-MM이 아니다"
      summary: 고객별 월 사용량 조회
      tags:
      - billable-metric-usage-controller
components:
  schemas:
...
    BillableMetricUsageCustomer:
      type: object
      properties:
        customer_id:
          type: string
          format: uuid
        customer_name:
          type: string
          description: 집계 시점의 고객 이름
        quantity:
          type: number
          description: 이벤트 properties의 값을 합산한 수량. 소수 자릿수를 그대로 싣는다
    BillableMetricUsageResponse:
      type: object
      properties:
        aggregation:
          type: string
        code:
          type: string
        customers:
          type: array
          items:
            $ref: "#/components/schemas/BillableMetricUsageCustomer"
        event_type:
          type: string
        name:
          type: string
        target_property:
          type: string
...
    DraftInvoiceCustomer:
      type: object
      properties:
        amount:
          type: integer
          format: int64
        customer_id:
          type: string
          format: uuid
        customer_name:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/DraftInvoiceLine"
    DraftInvoiceLine:
      type: object
      properties:
        amount:
          type: integer
          format: int64
        billable_metric_code:
          type: string
        quantity:
          type: number
        target_property:
          type: string
        unit_price:
          type: number
    DraftInvoiceResponse:
      type: object
      properties:
        calculated_at:
          type: string
          format: date-time
        customers:
          type: array
          items:
            $ref: "#/components/schemas/DraftInvoiceCustomer"
        month:
          type: string
        total_amount:
          type: integer
          format: int64
    ErrorResponse:
      type: object
      description: "오류 응답. code로 오류 종류를 구분하고, message는 code마다 하나인 한국어 문구다."
      properties:
        code:
          type: string
          description: 오류 종류. 오류별 처리는 이 값으로 분기한다.
          enum:
          - validation_error
          - malformed_request_body
          - unknown_organization
          - invalid_event
          - invalid_billable_metric
          - invalid_price_policy
          - customer_not_found
          - billable_metric_not_found
          - endpoint_not_found
          - method_not_allowed
          - response_type_not_acceptable
          - customer_has_events
          - billable_metric_already_exists
          - price_policy_already_exists
          - request_type_not_supported
          - internal_server_error
        errors:
          type: array
          description: 틀린 필드 목록. 400 응답에만 실린다.
          items:
            $ref: "#/components/schemas/FieldError"
        message:
          type: string
          description: code마다 하나인 한국어 문구. 예고 없이 바뀔 수 있으니 분기에 쓰지 않는다.
          example: 서버 내부 오류입니다
      required:
      - code
      - message
    EventResponse:
      type: object
      properties:
        customer_id:
          type: string
          format: uuid
        customer_name:
          type: string
        occurred_at:
          type: string
          format: date-time
        properties:
          type: object
          additionalProperties: {}
        received_at:
          type: string
          format: date-time
        transaction_id:
          type: string
        type:
          type: string
    FieldError:
      type: object
      description: 틀린 필드 하나
      properties:
        field:
          type: string
          description: 틀린 값을 보낸 자리의 이름. 요청에 쓴 이름과 같다.
          example: event_type
        message:
          type: string
          description: 이 자리의 값이 거절된 이유.
          example: 공백일 수 없습니다
    IngestEventRequest:
      type: object
      properties:
        customer_id:
          type: string
          format: uuid
        properties:
          type: object
          additionalProperties: {}
        timestamp:
          type: string
          format: date-time
        transaction_id:
          type: string
          maxLength: 255
          minLength: 0
        type:
          type: string
          minLength: 1
      required:
      - customer_id
      - properties
      - timestamp
      - transaction_id
      - type
    IngestEventResponse:
      type: object
      properties:
        duplicate:
          type: boolean
        transaction_id:
          type: string
...
    ListBillableMetricUsagesResponse:
      type: object
      properties:
        billable_metric_usages:
          type: array
          items:
            $ref: "#/components/schemas/BillableMetricUsageResponse"
        month:
          type: string
...
    ListEventsResponse:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/EventResponse"
        month:
          type: string
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
...