# backend/src/main/java/com/meterengine/invoice/controller/DraftInvoiceController.java
package com.meterengine.invoice.controller;

import com.meterengine.global.error.ErrorResponse;
import com.meterengine.invoice.dto.DraftInvoiceResponse;
import com.meterengine.invoice.service.DraftInvoiceService;
import com.meterengine.metric.service.BillableMetricUsageService;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import java.time.YearMonth;
import java.util.UUID;
import org.springframework.format.annotation.DateTimeFormat;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/v1/invoices/draft")
public class DraftInvoiceController {

  private final DraftInvoiceService draftInvoiceService;

  DraftInvoiceController(DraftInvoiceService draftInvoiceService) {
    this.draftInvoiceService = draftInvoiceService;
  }

  @GetMapping
  @Operation(
      summary = "청구 예정액 조회",
      description =
          """
          모든 고객의 청구 예정액을 미터별 사용량 x 단가로 계산해 돌려준다.
          조회할 때마다 계산하며 저장하지 않는다. 기간은 KST 기준의 달이다.
          금액은 원 단위 정수다. 라인(고객 x 미터)마다 원 미만을 버린 뒤 합산한다.
          이벤트가 없는 고객도 단가가 있는 미터마다 수량 0, 금액 0인 라인이 나온다.
          고객은 이름 오름차순, 라인은 미터 code 오름차순이다.
          """)
  @ApiResponses({
    @ApiResponse(responseCode = "200", description = "고객별 청구 예정액. 등록한 미터가 없으면 각 고객의 lines가 빈 배열이다"),
    @ApiResponse(
        responseCode = "400",
        content =
            @Content(
                mediaType = "application/json",
                schema = @Schema(implementation = ErrorResponse.class)),
        description =
            "code=validation_error: X-Organization-Id가 없거나 UUID가 아니거나, month가 yyyy-MM이 아니다")
  })
  public DraftInvoiceResponse previewDraftInvoice(
      @Parameter(description = "도입사 ID. 인증이 붙기 전까지 쓰는 임시 헤더다.") @RequestHeader("X-Organization-Id")
          UUID organizationId,
      @Parameter(description = "집계할 달(yyyy-MM, KST). 생략하면 이번 달이다.", example = "2026-08")
          @RequestParam(required = false)
          @DateTimeFormat(pattern = "yyyy-MM")
          YearMonth month) {
    YearMonth target = month == null ? BillableMetricUsageService.currentMonth() : month;
    return draftInvoiceService.preview(organizationId, target);
  }
}